#1987: changes after codereview : Merged Csv2MdTable into CsvTextConverter
