+++ app/src/main/java/net/gsantner/markor/format/csv/CsvConfig.java
@@ -14,8 +14,8 @@ package net.gsantner.markor.format.csv;
  * android and to Markor-Architecture.
  */
 public class CsvConfig {
-    public static final CsvConfig DEFAULT = new CsvConfig(';', '"');
-    public static final char[] CSV_DELIMITER_CANDIDATES = {DEFAULT.getFieldDelimiterChar(), ',', '\t', ':', '|'};
+    public static final CsvConfig DEFAULT = new CsvConfig(',', '"');
+    public static final char[] CSV_DELIMITER_CANDIDATES = {DEFAULT.getFieldDelimiterChar(), ';', '\t', ':', '|'};
     public static final char[] CSV_QUOTE_CANDIDATES = {DEFAULT.getQuoteChar(), '\''};
     private final char m_fieldDelimiterChar;
     private final char m_quoteChar;
